feat(ring_theory): induction and multiplicativity via prime factorization
Two pairs of lemmas for unique factorization domains, where we can show that a property holds on all elements, or a function is multiplicative, by showing this works on basecases, products of coprime elements and powers of prime elements.

We will use the last result to show the ideal norm is a multiplicative map.